[#217] refactor: 다녀온 여행지 리팩토링
