--- lib/common/button/image_button.dart
@@ -1,120 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:tripStory/common/button/base/base_button.dart';
-import 'package:tripStory/common/icon/svg_icon.dart';
-import 'package:tripStory/common/widget/round_thumbnail_image.dart';
-import 'package:tripStory/util/extension/context_extension.dart';
-
-class ImageButton extends StatelessWidget {
-  final XFile? pickedImage;
-  final String? url;
-  final VoidCallback? onPressed;
-  final String? errorIcon;
-  final double? errorIconSize;
-  final Color? iconBackgroundColor;
-  final Color? iconBorderColor;
-  final String iconPath;
-
-  const ImageButton({
-    super.key,
-    this.pickedImage,
-    this.onPressed,
-    this.url,
-    this.errorIcon,
-    this.errorIconSize,
-    this.iconBackgroundColor,
-    this.iconBorderColor,
-    required this.iconPath,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 98,
-      height: 98,
-      child: Stack(
-        children: [
-          BaseButton(
-            onTap: onPressed,
-            child: _buildImageContent(),
-          ),
-          Positioned(
-            right: 0,
-            bottom: 0,
-            child: Container(
-              width: 36,
-              height: 36,
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: iconBackgroundColor ?? context.color.white,
-                border: Border.all(
-                  width: 1,
-                  color: iconBorderColor ?? context.color.whiteEC,
-                ),
-              ),
-              child: BaseButton(
-                onTap: onPressed,
-                borderRadius: 100,
-                child: SvgIcon(
-                  assetPath: iconPath,
-                ),
-              ),
-            ),
-          ),
-          // Positioned(
-          //   bottom: 0,
-          //   right: 0,
-          //   child: BaseButton(
-          //     onTap: onPressed,
-          //     child: SvgPicture.asset("assets/icon/plus.svg"),
-          //   ),
-          // ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildImageContent() {
-    if (pickedImage != null) {
-      return ClipRRect(
-        borderRadius: BorderRadius.circular(4),
-        child: Image.file(
-          File(pickedImage!.path),
-          width: 80,
-          height: 80,
-          fit: BoxFit.fill,
-        ),
-      );
-    }
-    if (url != null) {
-      return RoundThumbnailImage(
-        imageUrl: url,
-        size: 80,
-        errorIcon: errorIcon,
-        errorIconSize: errorIconSize,
-      );
-    }
-
-    return Container(
-      width: 80,
-      height: 80,
-      decoration: BoxDecoration(
-        border: Border.all(color: const Color(0xFFEEEEEE), width: 1.0),
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Center(
-        child: SvgPicture.asset(
-          'assets/icon/image.svg',
-          width: 28,
-          colorFilter: const ColorFilter.mode(
-            Color(0xFFBDBDBD),
-            BlendMode.srcIn,
-          ), // gray400
-        ),
-      ),
-    );
-  }
-}
