+++ lib/common/image/cached_image.dart
@@ -1,8 +1,8 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:tripStory/common/image/empty_image.dart';
+import 'package:tripStory/common/icon/svg_icon.dart';
 import 'package:tripStory/core/constants/icon_constants.dart';
+import 'package:tripStory/util/extension/context_extension.dart';
 
 class CachedImage extends StatelessWidget {
   final String imageUrl;
@@ -26,22 +26,55 @@ class CachedImage extends StatelessWidget {
   Widget build(BuildContext context) {
     return CachedNetworkImage(
       imageUrl: imageUrl,
+      placeholder: (context, url) => SizedBox(
+        width: width,
+        height: height,
+      ),
       imageBuilder: (context, imageProvider) => Container(
         width: width,
         height: height,
         decoration: BoxDecoration(
-          image: DecorationImage(image: imageProvider, fit: fit),
+          image: DecorationImage(
+            image: imageProvider,
+            fit: fit,
+          ),
         ),
       ),
-      errorWidget: (context, url, error) => EmptyImage(
-        icon: SvgPicture.asset(
-          errorIcon ?? IconConstants.appLogo,
+      errorWidget: (context, url, error) => _EmptyImage(
+        icon: SvgIcon(
+          assetPath: errorIcon ?? IconConstants.appLogo,
           fit: BoxFit.fill,
           width: errorIconSize,
           height: errorIconSize,
         ),
         width: width,
         height: height,
+      ),
+    );
+  }
+}
+
+class _EmptyImage extends StatelessWidget {
+  final Widget icon;
+  final double width;
+  final double height;
+
+  const _EmptyImage({
+    required this.icon,
+    required this.width,
+    required this.height,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: width,
+      height: height,
+      decoration: BoxDecoration(
+        color: context.color.gray200,
+      ),
+      child: Center(
+        child: icon,
       ),
     );
   }
