+++ lib/domain/repositories/country_repository.dart
@@ -1,0 +1,6 @@
+import 'package:tripStory/core/network/typedefs.dart';
+import 'package:tripStory/domain/entities/visited_country_entity.dart';
+
+abstract class CountryRepository {
+  ResultFuture<List<VisitedCountryEntity>> fetchVisitedCountry();
+}
