+++ lib/view/hoom/controller/trip_rooms_create_controller.dart
@@ -9,10 +9,10 @@ import 'package:tripStory/data/models/trip_room_create_request.dart';
 import 'package:tripStory/domain/usecases/create_trip_room_usecase.dart';
 import 'package:tripStory/domain/usecases/fetch_presigned_url_usecase.dart';
 import 'package:tripStory/router/routes.dart';
-import 'package:tripStory/services/country_cache_manager.dart';
 import 'package:tripStory/util/compress_image.dart';
 import 'package:tripStory/util/extension/color_extension.dart';
 import 'package:tripStory/util/extension/date_extension.dart';
+import 'package:tripStory/util/helper/country_flag_helper.dart';
 import 'package:tripStory/util/helper/file_upload_helper.dart';
 import 'package:tripStory/util/one_time_event.dart';
 import 'package:tripStory/util/url_utils.dart';
@@ -35,7 +35,7 @@ class TripRoomsCreateController extends GetxController with GetSingleTickerProvi
   );
 
   Future<void> precacheFlags(BuildContext context) async {
-    await CountryFlagCacheManager.precacheAll(context);
+    await CountryFlagHelper.precacheAllRegions(context);
   }
 
   /// side Effect
